Updated VPP and code refactoring

+++ Backend/src/main/java/com/github/trionfettinicoUNICAM/PPTeam_DOIT/service/SimpleOrganizationsManager.java
@@ -1,6 +1,7 @@
 package com.github.trionfettinicoUNICAM.PPTeam_DOIT.service;
 
 import com.github.trionfettinicoUNICAM.PPTeam_DOIT.exception.EntityNotFoundException;
+import com.github.trionfettinicoUNICAM.PPTeam_DOIT.exception.IdConflictException;
 import com.github.trionfettinicoUNICAM.PPTeam_DOIT.model.*;
 import com.github.trionfettinicoUNICAM.PPTeam_DOIT.repository.OrganizationRepository;
 import com.github.trionfettinicoUNICAM.PPTeam_DOIT.repository.ProjectRepository;
@@ -9,11 +10,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpStatusCodeException;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 // TODO: 22/01/2021 controllare annotazioni perche forse service non e un singleton
@@ -27,20 +31,31 @@ public class SimpleOrganizationsManager implements OrganizationsManager{
     private ProjectRepository projectRepository;
 
     @Override
-    public Organization getOrganizationInstance(String organizationId) throws EntityNotFoundException {
+    public Organization getInstance(String organizationId) throws EntityNotFoundException {
+        if(organizationId.isBlank()) throw new IllegalArgumentException("Il campo 'ID' è vuoto");
         return organizationRepository.findById(organizationId).orElseThrow(()->
                 new EntityNotFoundException("Nessuna organizzazione trovata con l'ID: "+organizationId));
     }
 
     @Override
-    public Organization createNewOrganization(Organization organization) {
+    public Organization create(Organization organization) throws EntityNotFoundException, IdConflictException {
+        if(!userRepository.existsById(organization.getCreatorMail()))
+            throw new EntityNotFoundException("Nessun utente trovato con la mail: "+organization.getCreatorMail());
+        for (String memberMail : organization.getMembersMails()) {
+            if (!userRepository.existsById(memberMail))
+                throw new EntityNotFoundException("Nessun utente con la mail: "+memberMail);
+        }
+        if(existsName(organization.getName()))
+            throw new IdConflictException("Esiste gia un organizzazione con questo nome");
         if(!organization.getMembersMails().contains(organization.getCreatorMail()))
             organization.addMember(organization.getCreatorMail());
         return organizationRepository.save(organization);
     }
 
     @Override
-    public boolean deleteOrganization(String organizationId) {
+    public boolean delete(String organizationId) {
+        if(organizationId.isBlank()) throw new IllegalArgumentException("Il campo 'ID' è vuoto");
+        if(!exists(organizationId)) return false;
         organizationRepository.deleteById(organizationId);
         for (Project project: projectRepository.findByOrganizationId(organizationId)) {
             projectRepository.deleteById(project.getId());
@@ -49,12 +64,15 @@ public class SimpleOrganizationsManager implements OrganizationsManager{
     }
 
     @Override
-    public Organization updateOrganization(Organization organization) {
+    public Organization update(Organization organization) throws EntityNotFoundException {
+        if(!exists(organization.getId()))
+            throw new EntityNotFoundException("Nessuna organizzazione con id: "+ organization.getId());
         return organizationRepository.save(organization);
     }
 
     @Override
     public List<User> getUsers(String organizationId) throws EntityNotFoundException {
+        if(organizationId.isBlank()) throw new IllegalArgumentException("Il campo 'ID' è vuoto");
         Organization org = organizationRepository.findById(organizationId).orElseThrow(()->
                 new EntityNotFoundException("Nessuna organizzazione trovata con l'ID: "+organizationId));
         List<User> users = new ArrayList<>();
@@ -67,47 +85,101 @@ public class SimpleOrganizationsManager implements OrganizationsManager{
 
     @Override
     public boolean existsName(String organizationName) {
+        if(organizationName.isBlank()) throw new IllegalArgumentException("Il campo 'name' è vuoto");
         return organizationRepository.findAll().stream()
                 .anyMatch(it->it.getName().equals(organizationName));
     }
 
     @Override
     public boolean exists(String organizationId) {
+        if(organizationId.isBlank()) throw new IllegalArgumentException("Il campo 'ID' è vuoto");
         return organizationRepository.existsById(organizationId);
     }
 
     @Override
     public List<Organization> findByUser(String userMail) {
+        if(userMail.isBlank()) throw new IllegalArgumentException("Il campo 'userMail' è vuoto");
         return organizationRepository.findByMember(userMail);
     }
 
     @Override
     public List<Organization> findByCreator(String userMail) {
+        if(userMail.isBlank()) throw new IllegalArgumentException("Il campo 'userMail' è vuoto");
         return organizationRepository.findByCreatorMail(userMail);
     }
 
 
     @Override
-    public Page<BasicOrganizationInformation> getPage(int page, int size) {
+    public Page<String> getPage(int page, int size) {
         Page<Organization> projectPage = organizationRepository.findAll(PageRequest.of(page, size));
-        List<BasicOrganizationInformation> basicOrganizationInformation = new java.util.ArrayList<>(Collections.emptyList());
+        List<String> basicOrganizationInformation = new java.util.ArrayList<>(Collections.emptyList());
         for(Organization organization : projectPage){
-            User creator = userRepository.findById(organization.getCreatorMail()).orElseThrow(IllegalArgumentException::new);
-            basicOrganizationInformation.add(new BasicOrganizationInformation(organization,creator));
+            basicOrganizationInformation.add(getBasicJsonInformation(organization));
         }
         return new PageImpl<>(basicOrganizationInformation);
     }
 
     @Override
     public void addCollaborator(String organizationId, String userMail, Skill skill) throws EntityNotFoundException {
+        if(organizationId.isBlank()) throw new IllegalArgumentException("Il campo 'ID' è vuoto");
+        if(userMail.isBlank()) throw new IllegalArgumentException("Il campo 'userMail' è vuoto");
         Organization organization = organizationRepository.findById(organizationId).orElseThrow(()->
                 new EntityNotFoundException("Nessun organizzazione trovata con l'ID: "+organizationId));
-        organization.addMember(userMail);
-        organizationRepository.save(organization);
         User user = userRepository.findById(userMail).orElseThrow(()->
                 new EntityNotFoundException("Nessun utente trovato con la mail: "+userMail));
+        organization.addMember(userMail);
+        organizationRepository.save(organization);
         user.setExpert(skill,organizationId);
         userRepository.save(user);
     }
 
+    @Override
+    public boolean removeMember(String organizationId, String memberMail) throws EntityNotFoundException {
+        if(organizationId.isBlank()) throw new IllegalArgumentException("Il campo 'ID' è vuoto");
+        if(memberMail.isBlank()) throw new IllegalArgumentException("Il campo 'memberMail' è vuoto");
+        Organization organization = getInstance(organizationId);
+        User user = userRepository.findById(memberMail).orElseThrow(()->
+                new EntityNotFoundException("Nessun utente trovato con la mail: "+memberMail));
+        if(!organization.getMembersMails().contains(memberMail)) return false;
+        organization.removeMember(memberMail);
+        user.getSkills().forEach(skill -> skill.getExpertInOrganization().remove(organizationId));
+        User savedUser = userRepository.save(user);
+        Organization savedOrg = organizationRepository.save(organization);
+        return organization.equals(savedOrg) && user.equals(savedUser);
+    }
+
+    @Override
+    public boolean addMember(String organizationId, String memberMail) throws EntityNotFoundException {
+        if(organizationId.isBlank()) throw new IllegalArgumentException("Il campo 'ID' è vuoto");
+        if(memberMail.isBlank()) throw new IllegalArgumentException("Il campo 'memberMail' è vuoto");
+        Organization organization = getInstance(organizationId);
+        if(!userRepository.existsById(memberMail))
+            throw new EntityNotFoundException("Nessun utente trovato con la mail: "+memberMail);
+        if(organization.getMembersMails().contains(memberMail)) return false;
+        organization.addMember(memberMail);
+        Organization savedOrg = organizationRepository.save(organization);
+        return organization.equals(savedOrg);
+    }
+
+    @Override
+    public String getBasicJsonInformation(Organization object) {
+        StringBuilder builder = new StringBuilder();
+        builder.append("{");
+        builder.append("'id':");
+        builder.append("'" + object.getId() + "',");
+        builder.append("'name':");
+        builder.append("'" + object.getName() + "',");
+        builder.append("'description':");
+        builder.append("'" + object.getDescription() + "',");
+        builder.append("'creatorMail':");
+        builder.append("'" + object.getCreatorMail() + "',");
+        builder.append("'creatorName':");
+        Optional<User> optionalUser = userRepository.findById(object.getCreatorMail());
+        if(optionalUser.isPresent())
+            builder.append("'" + optionalUser.get().getName() + "'");
+        else
+            builder.append("'No user found'");
+        builder.append("}");
+        return builder.toString();
+    }
 }
